chore: change config and usage metric (#23)
Because

- CORS configuration is set in docker-compose files
- replace custom endpoint style `:` with `/`

This commit

- upgrade pre-commit hooks
- update action scripts
- refine .env file
- remove cors default config
- adopt latest protogen-go

+++ pkg/handler/handler.go
@@ -132,9 +132,9 @@ func (h *handler) ListUser(ctx context.Context, req *mgmtPB.ListUserRequest) (*m
 	}
 
 	resp := mgmtPB.ListUserResponse{
-		Users:         pbUsers,
+		Users: pbUsers,
 		NextPageToken: nextPageToken,
-		TotalSize:     int64(totalSize),
+		TotalSize: totalSize,
 	}
 	return &resp, nil
 }
